Menambahkan authorization

+++ routes/api.php
@@ -2,27 +2,28 @@
 
 use App\Http\Controllers\Api\AuthController;
 use App\Http\Controllers\Api\UserController;
+use App\Http\Middleware\RoleMiddleware;
 use Illuminate\Support\Facades\Route;
 
 Route::prefix('auth')->group(function () {
 
-    Route::post('/login', [AuthController::class, 'login']);
+    Route::post('/login', [AuthController::class, 'login'])->name('login');
 
     Route::middleware('auth:sanctum')->group(function () {
-        Route::get('/me', [AuthController::class, 'me']);
-        Route::post('/logout', [AuthController::class, 'logout']);
+        Route::get('/me', [AuthController::class, 'me'])->name('auth.me');
+        Route::post('/logout', [AuthController::class, 'logout'])->name('auth.logout');
     });
 
 });
 
-Route::middleware(['auth:sanctum', 'role:Admin'])
+Route::middleware(['auth:sanctum', RoleMiddleware::class . ':Admin'])
     ->prefix('users')
     ->group(function () {
 
         Route::get('/', [UserController::class, 'index']);
         Route::post('/', [UserController::class, 'store']);
-        Route::get('/{user}', [UserController::class, 'show']);
-        Route::put('/{user}', [UserController::class, 'update']);
-        Route::delete('/{user}', [UserController::class, 'destroy']);
+        Route::get('/{user:user_id}', [UserController::class, 'show']);
+        Route::put('/{user:user_id}', [UserController::class, 'update']);
+        Route::delete('/{user:user_id}', [UserController::class, 'destroy']);
 
     });
